rerfactor: split some utilities
